vrg: allow user to specify delay for checkhooks
We don't have a checkhook implementation yet. Instead of running the
check hook, we just delay the next group execution in the backup or
recover sequence.

The default delay is 5 seconds.

If the recipe has an annotation with the key
ramendr.openshift.io/check-hook-delay then the value of this annotation
will determine the delay in seconds.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>