velero: add metadata only label for kubevirt plugin
Adding velero.kubevirt.io/metadataBackup label to the backup request
ensures that kubevirt velero plugin skips the restore check and
consistency-specific checks when a backup is taken. This is required as
we backup the pvc/pv ourselves in Ramen and skip them from the velero
backup; otherwise the checks fail.

Co-Authored-by: Annaraya Narasagond <annarayanarasagond@gmail.com>
Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>